Move plugin to runtime
