+++ flashlight/app/asr/tools/benchmark/ArchBenchmark.cpp
@@ -12,7 +12,7 @@
 #include <glog/logging.h>
 
 #include "flashlight/ext/common/SequentialBuilder.h"
-#include "flashlight/ext/plugin/ModulePlugin.h"
+#include "flashlight/pkg/runtime/plugin/ModulePlugin.h"
 #include "flashlight/fl/flashlight.h"
 #include "flashlight/lib/common/String.h"
