+++ flashlight/ext/CMakeLists.txt
@@ -1,9 +1,8 @@
-cmake_minimum_required(VERSION 3.10)
+cmake_minimum_required(VERSION 3.13)
 
 # ------------------------- Components -------------------------
 
 include(${CMAKE_CURRENT_LIST_DIR}/common/CMakeLists.txt)
-include(${CMAKE_CURRENT_LIST_DIR}/plugin/CMakeLists.txt)
 
 cmake_dependent_option(FL_EXT_BUILD_INTEGRATIONS
   "Build and link Halide integration" ON
